% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp.R
\name{hang_out}
\alias{hang_out}
\title{This episode will feature the adventures of...}
\usage{
hang_out(.data, friends = 2)
}
\arguments{
\item{.data}{A table}

\item{friends}{Number of rows to randomly select. Defaults to 2.}
}
\description{
It chooses two or more randomly selected records.  The assumption is
that each record represents a single friend in the group.
}
\examples{
library(mylittlepony)
hang_out(my_little_pony)

}
