% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp.R
\name{magic_of_friendship}
\alias{magic_of_friendship}
\title{Witness the Magic of Friendship!!}
\usage{
magic_of_friendship(.data, name = "name")
}
\arguments{
\item{.data}{A table with two or more records to compare}

\item{name}{The variable that contains the proper name. Defaults to "name".}
}
\description{
A function that compares two or more friends and let them know
that things they have in common, and that things they can learn
from each other
}
\examples{
library(mylittlepony)
magic_of_friendship(my_little_pony[1:2, ])

}
