% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my_little_pony}
\alias{my_little_pony}
\title{Pony data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 6 columns.}
\usage{
my_little_pony
}
\description{
The package includes a small dataset with each of the main pony's data.
The data includes preferences and features that when compared to another
inside the group, they could be something they have in common, or something
they could learn from each other.
}
\details{
\describe{
  \item{name}{Pony's name}
  \item{kind}{Kind of pony (Alicorn, pony, pegasus or unicorn)}
  \item{canfly}{Is able to fly (TRUE/FALSE)}
}
}
\keyword{datasets}
